Rationalize integration test use traits and resetAppConfigs
